Handle more errors after rebaes

+++ ZAPDTR/ZAPD/ZResource.cpp
@@ -288,7 +288,7 @@ void ZResource::GetSourceOutputCode([[maybe_unused]] const std::string& prefix)
 			ZArray* arr = (ZArray*)this;
 			if (arr->resList[0]->GetResourceType() == ZResourceType::Vertex)
 			{
-				for (int i = 0; i < arr->resList.size(); i++)
+				for (size_t i = 0; i < arr->resList.size(); i++)
 				{
 					ZVtx* vtx = (ZVtx*)arr->resList[i];
 					decl->vertexHack.push_back(vtx);
